refactor(schema): renames galaxy => external
